Fix exportError and importPages
