+++ packages/inside/src/pages/auth/signUp/index.tsx
@@ -1,12 +1,8 @@
 import React from 'react'
-import { SignUp } from '@common/components'
+import { SignUpPage } from '@common/pages'
 
-const SignUpPage = () => {
-  return (
-    <div>
-      <SignUp />
-    </div>
-  )
+const SignUp = () => {
+  return <SignUpPage />
 }
 
-export default SignUpPage
+export default SignUp
