+++ packages/inside/src/pages/index.tsx
@@ -1,3 +1,5 @@
+import { HomePage } from '@common/pages'
+
 export default function Home() {
-  return <>Bigouel Inside</>
+  return <HomePage />
 }
