--- packages/inside/src/pages/main/home/index.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { Home } from '@common/pages'
-
-const HomePage = () => {
-  return (
-      <Home />
-  )
-}
-
-export default HomePage
