+++ shared/api/common/src/hooks/auth/usePostBbozzakSignUp.ts
@@ -5,7 +5,7 @@ import { authQueryKeys } from '../../libs/queryKeys'
 import { authUrl } from '../../libs/urlController'
 import { AxiosResponse } from 'axios'
 import { useSetRecoilState } from 'recoil'
-import { Page } from '../../../../../common'
+import { Page } from '../../../../../common/src/atoms'
 
 export const usePostSignUpBbozzak = () => {
   const setPage = useSetRecoilState(Page)
