+++ shared/api/common/src/hooks/auth/usePostProfessorSignUp.ts
@@ -3,7 +3,7 @@ import { post } from '../../libs/api/method'
 import { authQueryKeys } from '../../libs/queryKeys'
 import { authUrl } from '../../libs/urlController'
 import { useSetRecoilState } from 'recoil'
-import { Page } from '../../../../../common'
+import { Page } from '../../../../../common/src/atoms'
 import { AxiosResponse } from 'axios'
 
 export const usePostSignUpProfessor = () => {
@@ -25,7 +25,9 @@ export const usePostSignUpProfessor = () => {
     authQueryKeys.postSignUpProfessor(),
     (signUpValues) => post(authUrl.signUpPropessor(), signUpValues),
     {
-      onSuccess: (data) => {},
+      onSuccess: (data) => {
+        setPage(4)
+      },
       onError: (error) => {
         return console.log(error)
       },
