+++ shared/common/src/components/index.ts
@@ -1,6 +1,3 @@
-export { default as SignUpSuccess } from './SignUp'
-export { default as SignUp } from './SignUp'
-export { default as Login } from './Login'
 export { default as ValueInput } from './ValueInput'
 export { default as Header } from './Header'
 export { default as Sequence } from './Sequence'
