+++ shared/common/src/index.ts
@@ -1,3 +1,4 @@
+export * from './pages'
 export * from './components'
 export * from './layouts'
 export * from './styles'
