+++ shared/common/src/pages/index.ts
@@ -1,4 +1,6 @@
-export { default as Home } from './home'
-export { default as Lecture } from './lecture'
-export { default as Club } from './club'
-export { default as Notice } from './notice'
+export { default as LoginPage } from './login'
+export { default as SignUpPage } from './signUp'
+export { default as HomePage } from './home'
+export { default as LecturePage } from './lecture'
+export { default as ClubPage } from './club'
+export { default as NoticePage } from './notice'
