+++ shared/common/src/pages/notice/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React from 'react'
 import * as S from './style'
 import { Header } from '../../components'
 import Bg1 from '../../assets/png/mainBg1.png'
