+++ shared/common/src/pages/signUp/Pagination/Page1/style.ts
