+++ shared/common/src/pages/signUp/Pagination/Page2/index.tsx
@@ -1,4 +1,4 @@
-import ValueInput from '../../../ValueInput'
+import { ValueInput } from '../../../../components'
 import React, { ChangeEvent, useState } from 'react'
 import SignUpScrollContainer from '../SignUpScrollContainer'
 import SignUpButtonContainer from '../SignUpButtonContainer'
