+++ shared/common/src/pages/signUp/Pagination/SignUpSuccess/index.stories.tsx
